refactor: Move json from static to dynamic allocation
Using smart pointers we have moved json to heap. The json messages can be
just a handshake or can be huge files which will occupy lots of heap space
this is determental to the worker host performance. To prevent hogging memory
on host machine we have used smart pointers which handle allocation and
dealloctaion on thier own (ref count based). This also cleanups the json
lib usage.

Signed-off-by: Tejas Udupa S <tejasudupa1285@gmail.com>